Split Client into a BaseClient and the Client that will contain generated method from openAPI
